Basic type system implementation (#17)
* feat(error): FhY type error
* feat(utils): Partially ordered set and lattice implementations
* feat(type): Type lattice and promotion functions
* chore: Format changed files with ruff
* chore: Add/improve docstrings for type, lattice, and poset
* chore(poset): Ignore the type of the networkx import
* fix(symbol_table_builder): Type of shape param variables are not uint64
* chore: Addressed PR comments in type and lattice implementations
* chore: Update docstrings to reflect correct standards

+++ src/fhy/lang/ast/passes/symbol_table_builder.py
@@ -191,7 +191,7 @@ class SymbolTableBuilder(Visitor):
                     var_frame = ir.VariableSymbolTableFrame(
                         name=dimension,
                         type=ir.NumericalType(
-                            ir.PrimitiveDataType(ir.CoreDataType.INT)
+                            ir.PrimitiveDataType(ir.CoreDataType.UINT64)
                         ),
                         type_qualifier=ir.TypeQualifier.PARAM,
                     )
